Make user initiate the start of the game
In the web export, audio is blocked until the user interacts with the game. This was affecting the deal too, which waits for a sound effect to play before arranging the cards. Now the user has to click "New Game" to begin, which ensures a user interaction before the game starts.